enum設定をlib/domain/weather.dartへ移管
